Fix multiple disable/enable scroll on iOS device (#2)
When performing multiple disable/enable scroll on a row on an iOS device, the code was not working because :
- allTargetElements[targetElement] was basically doing targetElement.toString(), which resulted in something like `"[object HTMLDivElement]"`
- targetElement was nos removed when invoking `enableBodyScroll`

This fix uses an array, which stores `targetElement`s references to compare them easily.

Tested on an iOS device and works. Please let me know if you want me to change anything!

Also, thanks for this great lib.